+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Change log
 -----------
 
+* Add support for downgrades in resinhup and logging enhancements [Andrei]
 * Update default resinhup tag to v1.1.0 [Andrei]
 * Introduce meta-resin-morty for using meta-resin with poky morty [Florin]